+++ lib/knapsack_pro/runners/queue/cucumber_runner.rb
@@ -10,6 +10,7 @@ module KnapsackPro
           ENV['KNAPSACK_PRO_TEST_SUITE_TOKEN'] = KnapsackPro::Config::Env.test_suite_token_cucumber
           ENV['KNAPSACK_PRO_QUEUE_MODE_ENABLED'] = 'true'
           ENV['KNAPSACK_PRO_QUEUE_ID'] = KnapsackPro::Config::EnvGenerator.set_queue_id
+          ENV['KNAPSACK_PRO_CUCUMBER_OPTIONS'] = args.to_s
 
           adapter_class = KnapsackPro::Adapters::CucumberAdapter
           KnapsackPro::Config::Env.set_test_runner_adapter(adapter_class)